+++ monitoring/src/main/java/org/thingsboard/monitoring/util/ResourceUtils.java
@@ -15,6 +15,7 @@
  */
 package org.thingsboard.monitoring.util;
 
+import com.fasterxml.jackson.databind.JsonNode;
 import lombok.SneakyThrows;
 import org.thingsboard.common.util.JacksonUtil;
 
@@ -24,14 +25,21 @@ public class ResourceUtils {
 
     @SneakyThrows
     public static <T> T getResource(String path, Class<T> type) {
-        InputStream resource = ResourceUtils.class.getClassLoader().getResourceAsStream(path);
-        if (resource == null) {
-            throw new IllegalArgumentException("Resource not found for path " + path);
-        }
+        InputStream resource = getResourceStream(path);
         return JacksonUtil.OBJECT_MAPPER.readValue(resource, type);
     }
 
+    @SneakyThrows
+    public static JsonNode getResource(String path) {
+        InputStream resource = getResourceStream(path);
+        return JacksonUtil.OBJECT_MAPPER.readTree(resource);
+    }
+
     public static InputStream getResourceAsStream(String path) {
+        return getResourceStream(path);
+    }
+
+    private static InputStream getResourceStream(String path) {
         InputStream resource = ResourceUtils.class.getClassLoader().getResourceAsStream(path);
         if (resource == null) {
             throw new IllegalArgumentException("Resource not found for path " + path);
